wc_handler2/page2 to wc_handler/page
